+++ packages/babel-preset-udemy-website/tests/result/arrows.js
@@ -1,0 +1,5 @@
+"use strict";
+
+module.exports = function (it) {
+  return it;
+};