+++ packages/babel-preset-udemy-website/tests/result/spread.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = [].concat(babelHelpers.toConsumableArray('abcde'));