+++ packages/babel-preset-udemy-website/tests/result/template.js
@@ -1,0 +1,5 @@
+"use strict";
+
+var _templateObject = babelHelpers.taggedTemplateLiteralLoose(["1", "3"], ["1", "3"]);
+
+module.exports = String.raw(_templateObject, 2);