+++ packages/babel-preset-udemy-website/tests/source/dynamic-import.js
@@ -1,0 +1,5 @@
+function f() {
+    return import('./arrows').then(() => undefined);
+}
+
+f();