+++ packages/babel-preset-udemy-website/tests/source/for-of-destructuring.js
@@ -1,0 +1,6 @@
+const array = new Array(1e3).fill(1);
+let counter = 0;
+
+for (const [i, val] of array.entries()) counter += i + val;
+
+module.exports = counter;