+++ packages/babel-preset-udemy-website/tests/source/object-spread.js
@@ -1,0 +1,7 @@
+function React() {}
+function Component() {}
+
+module.exports = [
+    {a: 1, ...{b: 2}},
+    <Component a={ 1 } { ...{b: 2} } />
+];