+++ packages/babel-preset-udemy-website/tests/source/template.js
@@ -1,0 +1,1 @@
+module.exports = String.raw`1${2}3`;