+++ packages/eslint-plugin-udemy/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="4.0.2"></a>
+## [4.0.2](https://github.com/udemy/eslint-udemy/compare/eslint-plugin-udemy@4.0.1...eslint-plugin-udemy@4.0.2) (2018-02-07)
+
+
+
+
+**Note:** Version bump only for package eslint-plugin-udemy
+
 <a name="4.0.1"></a>
 ## [4.0.1](https://github.com/udemy/eslint-udemy/compare/eslint-plugin-udemy@4.0.0...eslint-plugin-udemy@4.0.1) (2017-12-07)